Crate rusoto_rds_data

Amazon RDS Data Service
Amazon RDS provides an HTTP endpoint to run SQL statements on an Amazon Aurora Serverless DB cluster. To run these statements, you work with the Data Service API.

For more information about the Data Service API, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

If you're using the service, you're probably looking for RdsDataClient and RdsData.

Structs

BatchExecuteStatementRequest
The request parameters represent the input of a SQL statement over an array of data.

BatchExecuteStatementResponse
The response elements represent the output of a SQL statement over an array of data.

BeginTransactionRequest
The request parameters represent the input of a request to start a SQL transaction.

BeginTransactionResponse
The response elements represent the output of a request to start a SQL transaction.

ColumnMetadata
Contains the metadata for a column.

CommitTransactionRequest
The request parameters represent the input of a commit transaction request.

CommitTransactionResponse
The response elements represent the output of a commit transaction request.

ExecuteSqlRequest
The request parameters represent the input of a request to run one or more SQL statements.

ExecuteSqlResponse
The response elements represent the output of a request to run one or more SQL statements.

ExecuteStatementRequest
The request parameters represent the input of a request to run a SQL statement against a database.

ExecuteStatementResponse
The response elements represent the output of a request to run a SQL statement against a database.

Field
Contains a value.

RdsDataClient
A client for the AWS RDS DataService API.

Record
A record returned by a call.

ResultFrame
The result set returned by a SQL statement.

ResultSetMetadata
The metadata of the result set returned by a SQL statement.

RollbackTransactionRequest
The request parameters represent the input of a request to perform a rollback of a transaction.

RollbackTransactionResponse
The response elements represent the output of a request to perform a rollback of a transaction.

SqlParameter
A parameter used in a SQL statement.

SqlStatementResult
The result of a SQL statement.

StructValue
A structure value returned by a call.

UpdateResult
The response elements represent the results of an update.

Value
Contains the value of a column.

Enums

BatchExecuteStatementError
Errors returned by BatchExecuteStatement

BeginTransactionError
Errors returned by BeginTransaction

CommitTransactionError
Errors returned by CommitTransaction

ExecuteSqlError
Errors returned by ExecuteSql

ExecuteStatementError
Errors returned by ExecuteStatement

RollbackTransactionError
Errors returned by RollbackTransaction

Traits

RdsData
Trait representing the capabilities of the AWS RDS DataService API. AWS RDS DataService clients implement this trait.