Trait rusoto_rds_data::RdsData

pub trait RdsData {
    fn batch_execute_statement(
        &self, 
        input: BatchExecuteStatementRequest
    ) -> RusotoFuture<BatchExecuteStatementResponse, BatchExecuteStatementError>;

    fn begin_transaction(
        &self, 
        input: BeginTransactionRequest
    ) -> RusotoFuture<BeginTransactionResponse, BeginTransactionError>;

    fn commit_transaction(
        &self, 
        input: CommitTransactionRequest
    ) -> RusotoFuture<CommitTransactionResponse, CommitTransactionError>;

    fn execute_sql(
        &self, 
        input: ExecuteSqlRequest
    ) -> RusotoFuture<ExecuteSqlResponse, ExecuteSqlError>;

    fn execute_statement(
        &self, 
        input: ExecuteStatementRequest
    ) -> RusotoFuture<ExecuteStatementResponse, ExecuteStatementError>;

    fn rollback_transaction(
        &self, 
        input: RollbackTransactionRequest
    ) -> RusotoFuture<RollbackTransactionResponse, RollbackTransactionError>;
}

Trait representing the capabilities of the AWS RDS DataService API. AWS RDS DataService clients implement this trait.

Required methods

fn batch_execute_statement(
&self,
input: BatchExecuteStatementRequest
) -> RusotoFuture<BatchExecuteStatementResponse, BatchExecuteStatementError>

Runs a batch SQL statement over an array of data.

    <p>You can run bulk update and insert operations for multiple records using a DML
statement with different parameter sets. Bulk operations can provide a significant
performance improvement over individual insert and update operations.</p>
<important>
<p>If a call isn't part of a transaction because it doesn't include the
<code>transactionID</code> parameter, changes that result from the call are
committed automatically.</p>
</important>

fn begin_transaction(
&self,
input: BeginTransactionRequest
) -> RusotoFuture<BeginTransactionResponse, BeginTransactionError>

Starts a SQL transaction.

    <important>
<p>A transaction can run for a maximum of 24 hours. A transaction is terminated and
rolled back automatically after 24 hours.</p>
<p>A transaction times out if no calls use its transaction ID in three minutes.
If a transaction times out before it's committed, it's rolled back
automatically.</p>
<p>DDL statements inside a transaction cause an implicit commit. We recommend
that you run each DDL statement in a separate <code>ExecuteStatement</code> call with
<code>continueAfterTimeout</code> enabled.</p>
</important>

fn commit_transaction(
&self,
input: CommitTransactionRequest
) -> RusotoFuture<CommitTransactionResponse, CommitTransactionError>

Ends a SQL transaction started with the BeginTransaction operation and commits the changes.

fn execute_sql(
&self,
input: ExecuteSqlRequest
) -> RusotoFuture<ExecuteSqlResponse, ExecuteSqlError>

Runs one or more SQL statements.

    <important>
<p>This operation is deprecated. Use the <code>BatchExecuteStatement</code> or
<code>ExecuteStatement</code> operation.</p>
</important>

fn execute_statement(
&self,
input: ExecuteStatementRequest
) -> RusotoFuture<ExecuteStatementResponse, ExecuteStatementError>

Runs a SQL statement against a database.

    <important>
<p>If a call isn't part of a transaction because it doesn't include the
<code>transactionID</code> parameter, changes that result from the call are
committed automatically.</p>
</important>
<p>The response size limit is 1 MB or 1,000 records. If the call returns more than 1 MB of response data or over 1,000 records, the call is terminated.</p>

fn rollback_transaction(
&self,
input: RollbackTransactionRequest
) -> RusotoFuture<RollbackTransactionResponse, RollbackTransactionError>

Performs a rollback of a transaction. Rolling back a transaction cancels its changes.

Loading content...